+++ src/qzxing/src/zxing/zxing/common/BitSource.cpp
@@ -71,6 +71,6 @@ int BitSource::readBits(int numBits) {
 }
 
 int BitSource::available() {
-  return 8 * (bytes_->size() - byteOffset_) - bitOffset_;
+  return 8 * (static_cast<int>(bytes_->size()) - byteOffset_) - bitOffset_;
 }
 }
